BUGFIX: UserData is uint64 - casting the ptr to uint64 (32bit env or 64bit env) prevents loss as opposed to uint (unpredictable size.)
git-svn-id: svn://svn.code.sf.net/p/nel/code/trunk@234 bf685bcc-7839-4fcd-9254-36823579bede

+++ nel/samples/pacs/object.cpp
@@ -117,7 +117,7 @@ CObjectDyn::CObjectDyn (double width, double depth, double height, double orient
 	_MovePrimitive->setAbsorbtion (0.9f);
 
 	// Setup user data
-	_MovePrimitive->UserData=(uint)this;
+	_MovePrimitive->UserData=(uint64)this;
 
 	// Setup trigger type
 	_MovePrimitive->setTriggerType (trigger);
@@ -195,7 +195,7 @@ CObjectDyn::CObjectDyn (double diameter, double height, const CVectorD& pos, con
 	_MovePrimitive->setAbsorbtion (0.9f);
 
 	// Setup user data
-	_MovePrimitive->UserData=(uint)this;
+	_MovePrimitive->UserData=(uint64)this;
 
 	// Setup trigger type
 	_MovePrimitive->setTriggerType (trigger);
